Use build tag so that iot agent doesn't get NCM.
